chore: remove Lumber Terminator util and rename TerminatorSender
