+++ src/services/schema/update/analyzer/sequelize-tables-analyzer.js
@@ -5,7 +5,7 @@ const ColumnTypeGetter = require('./sequelize-column-type-getter');
 const DefaultValueExpression = require('./sequelize-default-value');
 const TableConstraintsGetter = require('./sequelize-table-constraints-getter');
 const EmptyDatabaseError = require('../../../../errors/database/empty-database-error');
-const { terminate } = require('../../../../utils/terminator-sender');
+const { terminate } = require('../../../../utils/terminator');
 const stringUtils = require('../../../../utils/strings');
 const { isUnderscored } = require('../../../../utils/fields');
